Move internal windows structs and comment things
