+++ thcrap/src/plugin.cpp
@@ -209,11 +209,7 @@ static mod_funcs_t patch_funcs = {};
 static std::vector<HMODULE> plugins;
 
 TH_EXPORT void TH_CDECL export_jansson_funcs_mod_init(void*) {
-#ifdef _DEBUG
-	HMODULE jansson_handle = GetModuleHandleA("jansson_d.dll");
-#else
-	HMODULE jansson_handle = GetModuleHandleA("jansson.dll");
-#endif
+	HMODULE jansson_handle = GetModuleHandleA("jansson" DEBUG_OR_RELEASE ".dll");
 	if (exported_func_t* jansson_funcs = GetExportedFunctions(jansson_handle)) {
 		for (size_t i = 0; jansson_funcs[i].func != 0 && jansson_funcs[i].name != nullptr; i++) {
 			funcs[jansson_funcs[i].name] = jansson_funcs[i].func;
