+++ thcrap/src/util.h
@@ -40,210 +40,10 @@ confusing, so this macro exists to better document the intent.
 #define func_ptr_typedef(return_type, calling_convention, name) \
 typedef return_type (calling_convention* name)
 
+
 /// Internal Windows Structs
-/// TODO: Move these to some sort of header like windows_support.h
 /// -------
-
-/*
-Struct definitions based on the fields documented to have
-consistent offsets in all Windoes versions 5.0+
-TEB: https://www.geoffchappell.com/studies/windows/km/ntoskrnl/inc/api/pebteb/teb/index.htm
-PEB: https://www.geoffchappell.com/studies/windows/km/ntoskrnl/inc/api/pebteb/peb/index.htm
-*/
-
-typedef struct _PEB PEB;
-struct _PEB {
-	BOOLEAN InheritedAddressSpace;
-	BOOLEAN ReadImageFileExecOptions;
-	BOOLEAN BeingDebugged;
-	BOOLEAN SpareBool;
-#ifdef TH_X64
-	UCHAR Padding0[4];
-#endif
-	HANDLE Mutant;
-	PVOID ImageBaseAddress;
-	PVOID Ldr; // PEB_LDR_DATA*
-	PVOID ProcessParameters; // RTL_USER_PROCESS_PARAMETERS*
-	PVOID SubSystemData;
-	HANDLE ProcessHeap;
-	RTL_CRITICAL_SECTION* FastPebLock;
-	PVOID unreliable_member_1;
-	PVOID unreliable_member_2;
-	ULONG unreliable_member_3;
-#ifdef TH_X64
-	UCHAR Padding1[4];
-#endif
-	PVOID KernelCallbackTable;
-	ULONG SystemReserved[2];
-	PVOID unreliable_member_4;
-	ULONG TlsExpansionCounter;
-#ifdef TH_X64
-	UCHAR Padding2[4];
-#endif
-	PVOID TlsBitmap;
-	ULONG TlsBitmapBits[2];
-	PVOID ReadOnlySharedMemoryBase;
-	PVOID unreliable_member_5;
-	PVOID* ReadOnlyStaticServerData;
-	PVOID AnsiCodePageData;
-	PVOID OemCodePageData;
-	PVOID UnicodeCaseTableData;
-	ULONG NumberOfProcessors;
-	ULONG NtGlobalFlag;
-	LARGE_INTEGER CriticalSectionTimeout;
-	ULONG_PTR HeapSegmentReserve;
-	ULONG_PTR HeapSegmentCommit;
-	ULONG_PTR HeapDeCommitTotalFreeThreshold;
-	ULONG_PTR HeapDeCommitFreeBlockThreshold;
-	ULONG NumberOfHeaps;
-	ULONG MaximumNumberOfHeaps;
-	PVOID* ProcessHeaps;
-	PVOID GdiSharedHandleTable;
-	PVOID ProcessStarterHelper;
-	ULONG GdiDCAttributeList;
-#ifdef TH_X64
-	UCHAR Padding3[4];
-#endif
-	RTL_CRITICAL_SECTION* LoaderLock;
-	ULONG OSMajorVersion;
-	ULONG OSMinorVersion;
-	USHORT OSBuildNumber;
-	union {
-		USHORT OSCSDVersion;
-		struct {
-			BYTE OSCSDMajorVersion;
-			BYTE OSCSDMinorVersion;
-		};
-	};
-	ULONG OSPlatformId;
-	ULONG ImageSubsystem;
-	ULONG ImageSubsystemMajorVersion;
-	ULONG ImageSubsystemMinorVersion;
-#ifdef TH_X64
-	UCHAR Padding4[4];
-#endif
-	KAFFINITY unreliable_member_6;
-#ifdef TH_X64
-	ULONG GdiHandleBuffer[0x3C];
-#else
-	ULONG GdiHandleBuffer[0x22];
-#endif
-	VOID(*PostProcessInitRoutine)(VOID);
-	PVOID TlsExpansionBitmap;
-	ULONG TlsExpansionBitmapBits[0x20];
-	ULONG SessionId;
-#ifdef TH_X64
-	UCHAR Padding5[4];
-#endif
-};
-typedef struct _CLIENT_ID {
-	HANDLE UniqueProcess;
-	HANDLE UniqueThread;
-} CLIENT_ID;
-typedef struct _GDI_TEB_BATCH {
-	ULONG Offset;
-	ULONG_PTR HDC;
-	ULONG Buffer[310];
-} GDI_TEB_BATCH, *PGDI_TEB_BATCH;
-typedef struct _UNICODE_STRING {
-	USHORT Length;
-	USHORT MaximumLength;
-	PWSTR  Buffer;
-} UNICODE_STRING, *PUNICODE_STRING;
-typedef struct _TEB TEB;
-struct _TEB {
-	//NT_TIB NtTib;
-	struct _EXCEPTION_REGISTRATION_RECORD* ExceptionList;
-	PVOID StackBase;
-	PVOID StackLimit;
-	PVOID SubSystemTib;
-	PVOID FiberData;
-	PVOID ArbitraryUserPointer;
-	TEB* Self;
-	PVOID EnvironmentPointer;
-	CLIENT_ID ClientId;
-	PVOID ActiveRpcHandle;
-	PVOID ThreadLocalStoragePointer;
-	PEB* ProcessEnvironmentBlock;
-	ULONG LastErrorValue;
-	ULONG CountOfOwnedCriticalSections;
-	PVOID CsrClientThread;
-	PVOID Win32ThreadInfo;
-	ULONG User32Reserved[0x1A];
-	ULONG UserReserved[5];
-	PVOID WOW32Reserved;
-	ULONG CurrentLocale;
-	ULONG FpSoftwareStatusRegister;
-	PVOID SystemReserved1[0x36];
-	LONG ExceptionCode;
-#ifdef TH_X64
-	UCHAR Padding0[4];
-#endif
-	UCHAR SpareBytes1[0x2C];
-	GDI_TEB_BATCH GdiTebBatch;
-	CLIENT_ID RealClientId;
-	PVOID GdiCachedProcessHandle;
-	ULONG GdiClientPID;
-	ULONG GdiClientTID;
-	PVOID GdiThreadLocalInfo;
-	ULONG_PTR Win32ClientInfo[0x3E];
-	PVOID glDispatchTable[0xE9];
-	ULONG_PTR glReserved1[0x1D];
-	PVOID glReserved2;
-	PVOID glSectionInfo;
-	PVOID glSection;
-	PVOID glTable;
-	PVOID glCurrentRC;
-	PVOID glContext;
-	ULONG LastStatusValue;
-#ifdef TH_X64
-	UCHAR Padding2[4];
-#endif
-	UNICODE_STRING StaticUnicodeString;
-	union {
-		WCHAR StaticUnicodeBuffer[MAX_PATH + 1];
-		char StaticUTF8Buffer[(MAX_PATH + 1) * sizeof(WCHAR)];
-	};
-#ifdef TH_X64
-	UCHAR Padding3[6];
-#endif
-	PVOID DeallocationStack;
-	PVOID TlsSlots[0x40];
-	LIST_ENTRY TlsLinks;
-	PVOID Vdm;
-	PVOID ReservedForNtRpc;
-	HANDLE DbgSsReserved[2];
-};
-
-#ifdef TH_X64
-#define read_teb_member(member) (\
-member_size(TEB, member) == 1 ? read_gs_byte(offsetof(TEB, member)) : \
-member_size(TEB, member) == 2 ? read_gs_word(offsetof(TEB, member)) : \
-member_size(TEB, member) == 4 ? read_gs_dword(offsetof(TEB, member)) : \
-read_gs_qword(offsetof(TEB, member)) \
-)
-#define write_teb_member(member, data) (\
-member_size(TEB, member) == 1 ? write_gs_byte(offsetof(TEB, member), (data)) : \
-member_size(TEB, member) == 2 ? write_gs_word(offsetof(TEB, member), (data)) : \
-member_size(TEB, member) == 4 ? write_gs_dword(offsetof(TEB, member), (data)) : \
-write_gs_qword(offsetof(TEB, member), (data)) \
-)
-#else
-#define read_teb_member(member) (\
-member_size(TEB, member) == 1 ? read_fs_byte(offsetof(TEB, member)) : \
-member_size(TEB, member) == 2 ? read_fs_word(offsetof(TEB, member)) : \
-read_fs_dword(offsetof(TEB, member)) \
-)
-#define write_teb_member(member, data) (\
-member_size(TEB, member) == 1 ? write_fs_byte(offsetof(TEB, member), (data)) : \
-member_size(TEB, member) == 2 ? write_fs_word(offsetof(TEB, member), (data)) : \
-write_fs_dword(offsetof(TEB, member), (data)) \
-)
-#endif
-#define CurrentTeb() ((TEB*)read_teb_member(Self))
-#define CurrentPeb() ((PEB*)read_teb_member(ProcessEnvironmentBlock))
-
-#define KernelSharedDataAddr (0x7FFE0000u)
+#include "ntdll.h"
 
 #define CurrentImageBase ((uintptr_t)CurrentPeb()->ImageBaseAddress)
